arm64: rockchip_linux: Enable CONFIG_HZ_100.
As per Google, historically CONFIG_HZ is set to 100 on arm and arm64
devices, and that is not a good value for interactive devices.
However CONFIG_HZ_100 is perfectly fine for a device such as
the RK3588 or the RK3399 as they arent mobile devices which prioritize less jank.
Considering that we may wish to save and prefer energy efficiency
(users run several devices from the board) whilst running server like applications
CONFIG_HZ_100 seems more lucrative.

Bug: 149488822
Test: build pass and boot to home.
Signed-off-by: Jimmy Shiu <jimmyshiu@google.com>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>
Change-Id: I98c427e0fa95e4c4966516da52c049e58826e345
Signed-off-by: Rasenkai <rasenkai99@gmail.com>