phy: rockchip: inno-usb2: support phy pll control for rk3528
rk3528 doesn't has commonon control bit, previously we didn't
register the clk480m and use the disable variable of clkout_ctl
to distinguish whether clk480m registration is required. But
now we found that there is a phy inner debug register which can
control the pll out. So we can use this register to support phy
pll control for rk3528.

Signed-off-by: Jianwei Zheng <jianwei.zheng@rock-chips.com>
Change-Id: I86d5a4ba444b12dd0478836db34ec9980fbfed4e