drm/bridge: synopsys: Workaround for first-tone lost on TV-FFALCON
There are some TVs require maintaining N/CTS packets or AUDS
packets to keep audio logic active, otherwise, the first tone
may be lost.

In order to optimize the user experience, we need to ensure
continuous transmission of N/CTS and AUDS packets from the
HDMI-TX, so that the SINK TV devices can maintain audio logic
activation, promptly process audio data, and achieve the
completeness of the first tone.

So, we should not shutdown HDMI audio on the above situation.

This patch also update N/CTS to the correct value on BOOT-LOGO
situation to fix this issue.

Change-Id: Ie659a46bab63fb0f32b4a56b656713586bf795c7
Signed-off-by: Sugar Zhang <sugar.zhang@rock-chips.com>