Revert "rockchip: vop: add feature to disable lut for given vop"
This reverts commit ca99e9a90be77a691f13e21bbb4eb8d6d57e4197.

The extra lut disabling mechanism is not required as Xorg itself
will do if the connector is ignored.

In the /etc/X11/xorg.conf.d/x-modesetting.conf if the connector
has been ignored gamma lut will not be applied and still the
display can be used from the xserver, just RandR will not able
to see the display.