overlays: waveshare: define lane rate and swap x/y for ts
Without defining the lane rate, the calculated lane rate
was wrong for the display, so colors and displayed screen was
improper.

Setting it to the panel lane rate it fixes the issue with colors.

The goodix gt911 4.0inch panel requires to invert x and swap x/y
for the touchscreen.

Signed-off-by: djkabutar <d.kabutarwala@yahoo.com>